requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
redis>=5.0.0
groq>=0.3.0
httpx>=0.24.0
gunicorn>=21.0.0
pytest>=8.0.0
